requirements.txt
asgiref==3.6.0
coverage==7.1.0
Django==4.1.4
django-extensions==3.2.1
django-fsm==2.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Faker==16.8.1
pip-autoremove==0.10.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3